feat: claim logic is moved to useChainList Context - some tests are failing
